SCKAN-308 feat: Refactor export service
